requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.2
Django==5.2.4
django-filter==25.1
django-rest-framework==0.1.0
djangorestframework==3.16.0
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pydevd==3.3.0
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
